requirements.txt
anyio==3.5.0
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
email-validator==1.1.3
fastapi==0.75.0
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
orjson==3.6.7
pydantic==1.9.0
PyMySQL==1.0.2
python-dotenv==0.19.2
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.32
SQLAlchemy-Utils==0.38.2
starlette==0.17.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.2
iso8601==1.0.2
paho-mqtt==1.6.1
Werkzeug==2.0.3
